paging support for instruments
This is a trivial but important backward compatible fix to the `instruments` wrapper to allow for paging (by returning `'next'` field if no symbol passed in).  Paging for `instruments` endpoint is supported when no symbol is passed in, allowing the user to enumerate available instruments.  Existing users of this who pass in a symbol are unaffected.  The small change has been unit tested.